+++ elements/site-notice.php
@@ -1,0 +1,9 @@
+<div id="hw-scaffold-site-wide-notice" class="row">
+  <div class="container">
+    <div id="hw-scaffold-site-wide-notice-content">
+      <p>
+        <? echo get_theme_mod('scaffold_site_notice_text') ? get_theme_mod('scaffold_site_notice_text') : 'Set the "Site notice text"'; ?>
+      </p>
+    </div>
+  </div>
+</div>