+++ functions/functions-customiser.php
@@ -37,6 +37,20 @@ function scaffold_theme_customizer( $wp_customize ) {
 		'description' => 'Upload an alternate logo to replace the default site name and description',
 	) );
 
+  // Create a new section in the Theme Customizer
+  $wp_customize->add_section('scaffold_site_notice_section', array(
+    'title'       => __('Site notice', 'scaffold'),
+    'priority'    => 31,
+    'description' => 'Set a notice on every page',
+  ));
+  
+  // Create a new section in the Theme Customizer
+  $wp_customize->add_section('scaffold_demographic_section', array(
+    'title'       => __('Demographic Collection', 'scaffold'),
+    'priority'    => 32,
+    'description' => 'Configure demographic collection',
+  ));
+
 	// Create a new section in the Theme Customizer
 	$wp_customize->add_section( 'scaffold_org_identity_section' , array(
 		'title'       => __( 'Organisation Identity', 'scaffold' ),
@@ -54,6 +68,19 @@ function scaffold_theme_customizer( $wp_customize ) {
 	// Register the new setting
 	$wp_customize->add_setting( 'scaffold_logo' );
 	$wp_customize->add_setting( 'scaffold_logo_alt' );
+  // site notice settings
+  $wp_customize->add_setting('scaffold_site_notice_status', array(
+    'default'           => false
+  ));
+  $wp_customize->add_setting('scaffold_site_notice_text', array(
+    'default'           => __('', 'scaffold'),
+    'sanitize_callback' => 'sanitize_text'
+  ));
+  // demographic settings
+  $wp_customize->add_setting('scaffold_demographic_rate_review_url', array(
+    'default'           => __('', 'scaffold'),
+    'sanitize_callback' => 'sanitize_text'
+  ));
 	// org identity
 	$wp_customize->add_setting( 'scaffold_org_name', array(
 	 'default'           => __( '', 'scaffold' ),
@@ -72,6 +99,10 @@ function scaffold_theme_customizer( $wp_customize ) {
 	 'default'           => __( '', 'scaffold' ),
 	 'sanitize_callback' => 'sanitize_text'
 	));
+	$wp_customize->add_setting( 'scaffold_org_instagram', array(
+	 'default'           => __( '', 'scaffold' ),
+	 'sanitize_callback' => 'sanitize_text'
+	));
 	$wp_customize->add_setting( 'scaffold_org_twitter', array(
 	 'default'           => __( '', 'scaffold' ),
 	 'sanitize_callback' => 'sanitize_twitter'
@@ -81,6 +112,10 @@ function scaffold_theme_customizer( $wp_customize ) {
 	 'sanitize_callback' => 'sanitize_text'
 	));
 	$wp_customize->add_setting( 'scaffold_org_youtube', array(
+	 'default'           => __( '', 'scaffold' ),
+	 'sanitize_callback' => 'sanitize_text'
+	));
+	$wp_customize->add_setting( 'scaffold_org_mastodon', array(
 	 'default'           => __( '', 'scaffold' ),
 	 'sanitize_callback' => 'sanitize_text'
 	));
@@ -103,6 +138,44 @@ function scaffold_theme_customizer( $wp_customize ) {
 		'section'  => 'scaffold_logo_alt_section',
 		'settings' => 'scaffold_logo_alt',
 	) ) );
+
+  // Add site notice controls
+  $wp_customize->add_control(
+    new WP_Customize_Control($wp_customize, 'scaffold_site_notice_status', array(
+      'label'    => __('Show site notice?', 'scaffold'),
+      'section'  => 'scaffold_site_notice_section',
+      'settings' => 'scaffold_site_notice_status',
+      'type'     => 'checkbox'
+    ))
+  );
+  $wp_customize->add_control(new WP_Customize_Control(
+    $wp_customize,
+    'scaffold_site_notice_text',
+    array(
+      'label'    => __('Site notice text', 'scaffold'),
+      'section'  => 'scaffold_site_notice_section',
+      'settings' => 'scaffold_site_notice_text',
+      'type'     => 'text',
+      'input_attrs' => array(
+        'placeholder' => __('Your message here...', 'scaffold'),
+      )
+    )
+  ));
+  
+  // Add demographic settings controls
+  $wp_customize->add_control(new WP_Customize_Control(
+    $wp_customize,
+    'scaffold_demographic_rate_review_url',
+    array(
+      'label'    => __('Rate & Review Survey URL', 'scaffold'),
+      'section'  => 'scaffold_demographic_section',
+      'settings' => 'scaffold_demographic_rate_review_url',
+      'type'     => 'text',
+      'input_attrs' => array(
+        'placeholder' => __('https://www.smartsurvey.co.uk/s/', 'scaffold'),
+      )
+    )
+  ));
 
 	// Add org identity controls
 	$wp_customize->add_control( new WP_Customize_Control( $wp_customize, 'scaffold_org_name', array(
@@ -135,8 +208,15 @@ function scaffold_theme_customizer( $wp_customize ) {
 		'type'     => 'text',
 		'description' => 'Page name only, not full URL'
 	) ) );
+	$wp_customize->add_control( new WP_Customize_Control( $wp_customize, 'scaffold_org_instagram', array(
+		'label'    => __( 'Instagram Profile', 'scaffold' ),
+		'section'  => 'scaffold_org_social_media_section',
+		'settings' => 'scaffold_org_instagram',
+		'type'     => 'text',
+		'description' => 'Username/handle only, not full URL'
+	) ) );
 	$wp_customize->add_control( new WP_Customize_Control( $wp_customize, 'scaffold_org_twitter', array(
-		'label'    => __( 'Twitter', 'scaffold' ),
+		'label'    => __( 'Twitter/X', 'scaffold' ),
 		'section'  => 'scaffold_org_social_media_section',
 		'settings' => 'scaffold_org_twitter',
 		'type'     => 'text',
@@ -155,6 +235,13 @@ function scaffold_theme_customizer( $wp_customize ) {
 		'settings' => 'scaffold_org_youtube',
 		'type'     => 'text',
 		'description' => 'Channel ID only, not full URL'
+	) ) );
+	$wp_customize->add_control( new WP_Customize_Control( $wp_customize, 'scaffold_org_mastodon', array(
+		'label'    => __( 'Mastodon', 'scaffold' ),
+		'section'  => 'scaffold_org_social_media_section',
+		'settings' => 'scaffold_org_mastodon',
+		'type'     => 'url',
+		'description' => 'Full URL (instance and username)'
 	) ) );
 	$wp_customize->add_control( new WP_Customize_Control( $wp_customize, 'scaffold_org_rss', array(
 		'label'    => __( 'Show RSS feed link?', 'scaffold' ),